refactor: decompose viewpoint router
