+++ .github/workflows/tox.yml
@@ -19,7 +19,7 @@ jobs:
         lfs: 'true'
     - uses: actions/setup-python@v5
       with:
-        python-version: "3.10"
+        python-version: "3.11"
     - name: Run Tox
       run: |
         python -m pip install --upgrade pip tox
