+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.4.0
+    rev: v5.0.0
     hooks:
       - id: check-yaml
       - id: end-of-file-fixer
@@ -21,10 +21,10 @@ repos:
         types: [file, python]
 
   - repo: https://github.com/PyCQA/flake8
-    rev: 6.0.0
+    rev: 7.1.1
     hooks:
       - id: flake8
-        args: ["--ignore=E203,W503,W605", "--max-line-length=125"]
+        args: ["--max-line-length=125", "--max-complexity=10"]
         types: [file, python]
 
   - repo: https://github.com/pre-commit/mirrors-autopep8
@@ -40,7 +40,7 @@ repos:
         types: [file, python]
 
   - repo: https://github.com/aws-samples/automated-security-helper
-    rev: '1.0.9-e-16May2023' # update with the latest-tagged version in the repository
+    rev: 'v1.5.1'
     hooks:
         - id: ash
           stages: [ manual ]
