+++ setup.cfg
@@ -23,18 +23,18 @@ zip_safe = False
 package_dir=
     =src
 packages=find_namespace:
-python_requires = >=3.9
+python_requires = >=3.11
 include_package_data = True
 
 install_requires =
-    uvicorn
-    fastapi
-    fastapi-versioning
-    asgi-correlation-id
-    cryptography
-    boto3
-    geojson>=3.0.0
-    python-json-logger>=2.0.0
+    uvicorn==0.31.*
+    fastapi==0.115.*
+    fastapi-versioning==0.10.*
+    asgi-correlation-id==4.3.*
+    cryptography==43.0.*
+    boto3==1.35.*
+    geojson==3.1.*
+    python-json-logger==2.0.*
     osml-imagery-toolkit>=1.4.0rc4
 
 [options.packages.find]
