+++ src/aws/osml/tile_server/services/aws_services.py
@@ -1,17 +1,24 @@
 #  Copyright 2023-2024 Amazon.com, Inc. or its affiliates.
 
 import logging
+import sys
 import traceback
 from datetime import datetime, timezone
 from time import time
-from typing import Dict, Tuple
+from typing import Dict
 
 from boto3 import Session
 from boto3.resources.base import ServiceResource
 from botocore.credentials import RefreshableCredentials
+from botocore.exceptions import ClientError
 from botocore.session import get_session
 
 from aws.osml.tile_server.app_config import BotoConfig, ServerConfig
+
+from .database import ViewpointStatusTable
+from .queue import ViewpointRequestQueue
+
+logger = logging.getLogger("uvicorn")
 
 
 class RefreshableBotoSession:
@@ -23,7 +30,6 @@ class RefreshableBotoSession:
     def __init__(self):
         self.sts_arn = ServerConfig.sts_arn
         self.session_name = "TileServerSession"
-        self.logger = logging.getLogger("uvicorn")
         self.session_duration = 3600  # 60 minutes
 
     def _refresh(self) -> Dict[str, str]:
@@ -64,7 +70,7 @@ class RefreshableBotoSession:
                     "token": None,
                     "expiry_time": datetime.fromtimestamp(time() + 3600, timezone.utc).isoformat(),
                 }
-                self.logger.error(f"Error has occurred when setting up STS session - {err} / {traceback.format_exc()}")
+                logger.error(f"Error has occurred when setting up STS session - {err} / {traceback.format_exc()}")
 
         return credentials
 
@@ -88,55 +94,66 @@ class RefreshableBotoSession:
         return Session(botocore_session=session)
 
 
-def initialize_ddb(session: Session) -> ServiceResource:
-    """
-    Initialize DynamoDB service and return a service resource.
+class AwsServices:
+    def __init__(self, ddb: ServiceResource = None, s3: ServiceResource = None, sqs: ServiceResource = None) -> None:
+        """
+        Initialize AWS services required by the application.
 
-    :param: session: The credential session to use for the ServiceResource.
-    :return: DynamoDB service resource for consumption.
-    """
+        This function initializes DynamoDB, S3, and SQS services,
+        handling any exceptions that occur during the process.
 
-    return session.resource("dynamodb", config=BotoConfig.default, region_name=ServerConfig.aws_region)
+        :param: ddb: An optional DynamoDB service resource to use.  If none is provided a new one will be initialized.
+        :param: s3: An optional S3 service resource to use.  If none is provided a new one will be initialized.
+        :param: sqs: An optional SQS service resource to use.  If none is provided a new one will be initialized.
+        :return: None.
+
+        :raises: SystemExit if any service fails to initialize.
+        """
+        try:
+            session = RefreshableBotoSession().refreshable_session()
+            self.ddb = self.initialize_ddb(session) if ddb is None else ddb
+            self.s3 = self.initialize_s3(session) if s3 is None else s3
+            self.sqs = self.initialize_sqs(session) if sqs is None else sqs
+            if self.ddb is not None:
+                self.viewpoint_database = ViewpointStatusTable(self.ddb, logger)
+            if self.sqs is not None:
+                self.viewpoint_queue = ViewpointRequestQueue(self.sqs, ServerConfig.viewpoint_request_queue, logger)
+        except ClientError as err:
+            sys.exit(f"Fatal error occurred while initializing AWS services. Exiting. {err}")
+
+    @staticmethod
+    def initialize_ddb(session: Session) -> ServiceResource:
+        """
+        Initialize DynamoDB service and return a service resource.
+
+        :param: session: The credential session to use for the ServiceResource.
+        :return: DynamoDB service resource for consumption.
+        """
+
+        return session.resource("dynamodb", config=BotoConfig.default, region_name=ServerConfig.aws_region)
+
+    @staticmethod
+    def initialize_s3(session: Session) -> ServiceResource:
+        """
+        Initialize S3 service and return a service resource.
+
+        :param: session: The credential session to use for the ServiceResource.
+        :return: S3 service resource for consumption.
+        """
+
+        return session.resource("s3", config=BotoConfig.default)
+
+    @staticmethod
+    def initialize_sqs(session: Session) -> ServiceResource:
+        """
+        Initialize SQS service and return a service resource.
+
+        :param: session: The credential session to use for the ServiceResource.
+        :return: SQS service resource for consumption.
+        """
+
+        return session.resource("sqs", config=BotoConfig.default)
 
 
-def initialize_s3(session: Session) -> ServiceResource:
-    """
-    Initialize S3 service and return a service resource.
-
-    :param: session: The credential session to use for the ServiceResource.
-    :return: S3 service resource for consumption.
-    """
-
-    return session.resource("s3", config=BotoConfig.default)
-
-
-def initialize_sqs(session: Session) -> ServiceResource:
-    """
-    Initialize SQS service and return a service resource.
-
-    :param: session: The credential session to use for the ServiceResource.
-    :return: SQS service resource for consumption.
-    """
-
-    return session.resource("sqs", config=BotoConfig.default)
-
-
-def initialize_aws_services() -> Tuple[ServiceResource, ServiceResource, ServiceResource]:
-    """
-    Initialize AWS services required by the application.
-
-    This function initializes DynamoDB, S3, and SQS services,
-    handling any exceptions that occur during the process.
-
-    :return: Tuple containing the initialized service resources for ddb, s3, and sqs clients.
-
-    :raises: SystemExit if any service fails to initialize.
-    """
-
-    session = RefreshableBotoSession().refreshable_session()
-
-    ddb = initialize_ddb(session)
-    s3 = initialize_s3(session)
-    sqs = initialize_sqs(session)
-
-    return ddb, s3, sqs
+def get_aws_services() -> AwsServices:
+    return AwsServices()
