+++ src/aws/osml/tile_server/utils/__init__.py
@@ -1,7 +1,6 @@
 #  Copyright 2023-2024 Amazon.com, Inc. or its affiliates.
 
 # flake8: noqa
-from .aws_services import RefreshableBotoSession, initialize_aws_services, initialize_ddb, initialize_s3, initialize_sqs
 from .health_check import HealthCheck
 from .log_tools import ThreadingLocalContextFilter, configure_logger
 from .string_enums import AutoLowerStringEnum, AutoStringEnum, AutoUnderscoreStringEnum
@@ -12,4 +11,3 @@ from .tile_server_utils import (
     get_tile_factory_pool,
     perform_gdal_translation,
 )
-from .token import initialize_token_key, read_token_key
