+++ src/aws/osml/tile_server/viewpoint/__init__.py
@@ -1,15 +1,6 @@
 #  Copyright 2023-2024 Amazon.com, Inc. or its affiliates.
 
 # flake8: noqa
-from .database import DecimalEncoder, ViewpointStatusTable
-from .models import (
-    ViewpointApiNames,
-    ViewpointListResponse,
-    ViewpointModel,
-    ViewpointRequest,
-    ViewpointStatus,
-    ViewpointUpdate,
-)
-from .queue import ViewpointRequestQueue
-from .routers import ViewpointRouter
+
+from .router import viewpoint_router
 from .worker import SupplementaryFileType, ViewpointWorker
