+++ src/aws/osml/tile_server/viewpoint/viewpoint_id/__init__.py
@@ -1,0 +1,1 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
