+++ test/aws/osml/tile_server/viewpoint/test_create_viewpoint.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestCreateViewpoint(TestCase):
+    """Unit tests for create viewpoint endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_create_viewpoint(self):
+        pass
