+++ test/aws/osml/tile_server/viewpoint/test_list_viewpoints.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestListViewpoints(TestCase):
+    """Unit tests for list viewpoints endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_list_viewpoints(self):
+        pass
