+++ test/aws/osml/tile_server/viewpoint/test_update_viewpoint.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestUpdateViewpoint(TestCase):
+    """Unit tests for update viewpoint endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_update_viewpoint(self):
+        pass
