+++ test/aws/osml/tile_server/viewpoint/viewpoint_id/image/test_metadata.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestMetadata(TestCase):
+    """Unit tests for metadata endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_get_metadata(self):
+        pass
