+++ test/aws/osml/tile_server/viewpoint/viewpoint_id/image/test_preview.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestPreview(TestCase):
+    """Unit tests for preview endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_get_preview(self):
+        pass
