+++ test/aws/osml/tile_server/viewpoint/viewpoint_id/image/test_statistics.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestStatistics(TestCase):
+    """Unit tests for statistics endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_get_statistics(self):
+        pass
