+++ test/aws/osml/tile_server/viewpoint/viewpoint_id/map/test_tileset_metadata.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestMapTileset(TestCase):
+    """Unit tests for map tileset metadata endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_get_map_tileset_metadata(self):
+        pass
