+++ test/aws/osml/tile_server/viewpoint/viewpoint_id/map/test_tilesets.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestMapTilesets(TestCase):
+    """Unit tests for map tilesets endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_get_map_tilesets(self):
+        pass
