+++ test/aws/osml/tile_server/viewpoint/viewpoint_id/test_delete_viewpoint.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestDeleteViewpoint(TestCase):
+    """Unit tests for delete viewpoint endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_delete_viewpoint(self):
+        pass
