+++ test/aws/osml/tile_server/viewpoint/viewpoint_id/test_describe_viewpoint.py
@@ -1,0 +1,13 @@
+#  Copyright 2024 Amazon.com, Inc. or its affiliates.
+
+from unittest import TestCase
+
+import pytest
+
+
+class TestDescribeViewpoint(TestCase):
+    """Unit tests for describe viewpoint endpoint in tile_server."""
+
+    @pytest.mark.skip(reason="Test not implemented")
+    def test_describe_viewpoint(self):
+        pass
