+++ tox.ini
@@ -5,7 +5,7 @@
 [tox]
 envlist =
 # Basic configurations: Run the tests for each python version.
-    py{311}-prod
+    py{311}-prod, py{312}-prod
 
 # Build and test the docs with sphinx.
     docs
@@ -39,7 +39,7 @@ setenv =
     AWS_SECRET_ACCESS_KEY=testing
     AWS_SECURITY_TOKEN=testing
     AWS_SESSION_TOKEN=testing
-    EFS_MOUNT_NAME=/tmp/local_viewpoint_cache
+    EFS_MOUNT_NAME=tmp/local_viewpoint_cache
 commands =
     pytest --cov-config .coveragerc --cov aws.osml.tile_server --cov-report term-missing {posargs}
     {env:IGNORE_COVERAGE:} coverage html --rcfile .coveragerc
